requirements.txt
alembic==1.13.2
amqp==5.1.1
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.3.3
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-ftp==3.3.0
apache-airflow-providers-http==4.1.1
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-postgres==5.11.2
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.3.1
apispec==6.6.1
appnope==0.1.4
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.2.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==24.4.2
bleach==6.1.0
blinker==1.8.2
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
contourpy==1.2.1
cramjam==2.8.3
cron_descriptor==1.2.32
croniter==2.0.5
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.8
dnspython==2.3.0
docutils==0.19
email-validator==1.3.0
executing==2.0.1
fastjsonschema==2.19.1
fastparquet==2024.5.0
Flask==2.2.2
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.7.0
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.5.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
graphviz==0.20.1
grpcio==1.64.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.4.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.2.4
lazy-object-proxy==1.9.0
limits==3.13.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==5.2.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
methodtools==0.4.7
mistune==3.0.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.1.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.6.0
pylint==3.2.5
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==7.0.0
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.1.0
rich-argparse==1.5.2
rpds-py==0.18.1
Send2Trash==1.8.3
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.2.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tomlkit==0.12.5
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.4.0
tzdata==2024.1
uc-micro-py==1.0.1
unicodecsv==0.14.1
universal_pathlib==0.2.2
uri-template==1.3.0
urllib3==2.2.1
vine==5.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.11
wirerope==0.4.7
wrapt==1.14.1
WTForms==3.0.1
zipp==3.19.2
